refactor(core): move DeepReview runtime policy owner
